+++ apps/omsorgspengesoknad/src/app/pages/kvittering/KvitteringPage.tsx
@@ -1,7 +1,6 @@
 import { Heading, Link } from '@navikt/ds-react';
 import { useEffect } from 'react';
 import { useIntl } from 'react-intl';
-import { SIFCommonPageKey, useLogSidevisning } from '@navikt/sif-common-amplitude';
 import Block from '@navikt/sif-common-core-ds/src/atoms/block/Block';
 import CheckmarkIcon from '@navikt/sif-common-core-ds/src/atoms/checkmark-icon/CheckmarkIcon';
 import Checklist from '@navikt/sif-common-core-ds/src/components/lists/checklist/Checklist';
@@ -21,8 +20,6 @@ const KvitteringPage = ({ onUnmount }: Props) => {
             onUnmount();
         };
     });
-
-    useLogSidevisning(SIFCommonPageKey.kvittering);
 
     return (
         <Page title={text('page.kvittering.sidetittel')}>