+++ apps/pleiepenger-i-livets-sluttfase-soknad/src/app/pages/kvittering/KvitteringPage.tsx
@@ -1,5 +1,4 @@
 import { Heading, Link } from '@navikt/ds-react';
-import { SIFCommonPageKey, useLogSidevisning } from '@navikt/sif-common-amplitude';
 import Block from '@navikt/sif-common-core-ds/src/atoms/block/Block';
 import CheckmarkIcon from '@navikt/sif-common-core-ds/src/atoms/checkmark-icon/CheckmarkIcon';
 import Checklist from '@navikt/sif-common-core-ds/src/components/lists/checklist/Checklist';
@@ -22,8 +21,6 @@ const KvitteringPage = ({ kvitteringInfo, onUnmount }: Props) => {
             onUnmount();
         };
     });
-
-    useLogSidevisning(SIFCommonPageKey.kvittering);
 
     return (
         <Page title={text('page.kvittering.sidetittel')}>