+++ packages/sif-common-amplitude/src/index.ts
@@ -1,2 +1,1 @@
 export * from './amplitude';
-export { default as useLogSidevisning } from './hooks/useLogSidevisning';