+++ apps/omsorgspengesoknad/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @navikt/omsorgspengesoknad
 
+## 1.28.19
+
+### Patch Changes
+
+-   664f1eb: Fjerne egen logging av sidevisning til amplitude. Dette er erstattet av dekoratørens besøk event.
+-   Updated dependencies [664f1eb]
+    -   @navikt/sif-common-amplitude@2.25.16
+
 ## 1.28.18
 
 ### Patch Changes